optimize request for countVote

+++ src/main/java/com/rlrio/voting/controller/dto/vote/CountVoteDto.java
@@ -5,5 +5,5 @@ import lombok.Data;
 @Data
 public class CountVoteDto {
     private Long restaurantId;
-    private int votes;
+    private Long votes;
 }
